+++ internal/client.go
@@ -6,6 +6,7 @@ package internal
 import (
 	gomock "github.com/golang/mock/gomock"
 	fosite "github.com/ory/fosite"
+	go_jose "gopkg.in/square/go-jose.v2"
 )
 
 // Mock of Client interface
@@ -57,6 +58,26 @@ func (_m *MockClient) GetID() string {
 
 func (_mr *_MockClientRecorder) GetID() *gomock.Call {
 	return _mr.mock.ctrl.RecordCall(_mr.mock, "GetID")
+}
+
+func (_m *MockClient) GetJSONWebKeys() *go_jose.JSONWebKeySet {
+	ret := _m.ctrl.Call(_m, "GetJSONWebKeys")
+	ret0, _ := ret[0].(*go_jose.JSONWebKeySet)
+	return ret0
+}
+
+func (_mr *_MockClientRecorder) GetJSONWebKeys() *gomock.Call {
+	return _mr.mock.ctrl.RecordCall(_mr.mock, "GetJSONWebKeys")
+}
+
+func (_m *MockClient) GetJSONWebKeysURI() string {
+	ret := _m.ctrl.Call(_m, "GetJSONWebKeysURI")
+	ret0, _ := ret[0].(string)
+	return ret0
+}
+
+func (_mr *_MockClientRecorder) GetJSONWebKeysURI() *gomock.Call {
+	return _mr.mock.ctrl.RecordCall(_mr.mock, "GetJSONWebKeysURI")
 }
 
 func (_m *MockClient) GetRedirectURIs() []string {